Clean index.jsx
