+++ lib/actions/targetActions.js
@@ -34,7 +34,9 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { CommunicationType } from '../util/devices';
+
+import { logger } from 'nrfconnect/core';
+import { CommunicationType, VendorId, USBProductIds } from '../util/devices';
 import * as jlinkTargetActions from './jlinkTargetActions';
 import * as usbsdfuTargetActions from './usbsdfuTargetActions';
 
@@ -110,6 +112,30 @@ export function writeProgressFinishedAction() {
 export function writeProgressStartAction() {
     return {
         type: WRITE_PROGRESS_START,
+    };
+}
+
+export function openDevice(device) {
+    return dispatch => {
+        const serialNumber = device.serialNumber;
+        if (device.traits.includes('jlink')) {
+            dispatch(jlinkTargetActions.loadDeviceInfo(serialNumber));
+            return;
+        }
+        if (device.traits.includes('nordicUsb')) {
+            dispatch(usbsdfuTargetActions.openDevice(device));
+            return;
+        }
+        if (device.traits.includes('serialport')) {
+            const { vendorId, productId } = device.serialport;
+            const vid = parseInt(vendorId.toString(16), 16);
+            const pid = parseInt(productId.toString(16), 16);
+            if (vid === VendorId.NORDIC_SEMICONDUCTOR && USBProductIds.includes(pid)) {
+                dispatch(usbsdfuTargetActions.openDevice(device));
+                return;
+            }
+        }
+        logger.error('Unsupported device. The detected device is neither JLink device nor Nordic USB device.');
     };
 }
